批量关闭的确认目标改存页 id，不再误清别页
问题：三页 [A 置脏, B 干净, C 置脏(活动)] → 在 C 上「关闭其他」弹聚合确认
（目标 A+B）→ 确认条停留期间点 B 的 × 把 B 关掉 → 此时〔放弃并关闭〕清掉并
移除的是 **A 和 C**：C 的 dirty 被清、path 被丢、缓冲区被掏空后整页消失。用户
的未保存内容无声损毁，而且因为置脏标记是我们自己清的，连一次确认都不弹。视图
侧「N 个中有 M 个未保存」同样按陈旧下标现查，数字也跟着说谎。

原因：batch_close_confirm 存的是裸下标，而 close_tab_now 只对 close_tab_confirm
做了左移平移（P145）——同族的批量列表漏在规矩外面。

改法：列表改存 TabId（`begin_batch_close` 处一次性取出），消费与绘制都按 id
现地解析，解析不到 = 期间已被关掉，跳过。与本仓既成规矩对齐：`Saved(tab_id,
version)`、`Loaded` 按 id 定位目标页。本轮只动批量这一条；审计发现的另两条同族
持有者（renaming_tab 存下标、external_change 存下标）尚未一并改，见台账 §5。

验证：全量 768 绿、clippy --all-targets -D warnings exit=0。新增
batch_close_confirm_survives_a_tab_closed_while_it_is_shown 钉住「条子停留期间
关别页 → 只有真目标被清被移，幸存页的正文与置脏原样不动」；把两处改回旧形状
（按下标存 + 按下标动手）后它精确转红（left Some([0, 1]) vs right Some([1, 2])），
复原后绿。既有两条批量用例的断言由「等于下标」改成「等于该页 id」——顺带说明
它们此前测的是编码而不是行为。

+++ crates/app/src/state.rs
@@ -256,9 +256,12 @@ pub(crate) struct Editpad {
     pub(crate) cursor_pos: (f32, f32),
     /// 窗口逻辑尺寸（Resized 事件；(0,0) = 未知，贴边钳制跳过）
     pub(crate) viewport_size: (f32, f32),
-    /// Some(targets) = 批量关闭（关闭其他/右侧）目标列表，任一置脏时
-    /// 先弹一次聚合确认；确认后统一放弃并移除。固定页不在列表内。
-    pub(crate) batch_close_confirm: Option<Vec<usize>>,
+    /// Some(targets) = 批量关闭（关闭其他/右侧）目标**页 id** 列表，任一置脏
+    /// 时先弹一次聚合确认；确认后统一放弃并移除。固定页不在列表内。
+    /// P213：曾是裸下标——确认条停留期间关掉任意一页，整表就指到别的页上
+    /// （「放弃并关闭」会把用户没选过的页清空并移除）。跨异步/跨帧持有页
+    /// 引用一律用 id，与本仓 `Saved(tab_id, ..)` 同规矩。
+    pub(crate) batch_close_confirm: Option<Vec<u64>>,
     // ---------- 未命名页编号（P25） ----------
     /// 下一个未命名页序号（P25 曾为全局单调赋号的来源；P168 起赋号改
     /// 走「最小空闲复用」（见 `assign_untitled_num`），本字段不再被
